+++ src/main/java/duke/command/CommandsTypes.java
@@ -1,3 +1,5 @@
+package duke.command;
+
 public enum CommandsTypes {
     Exit,
     Add,
