+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.FileManager;
+import duke.Tasklist;
+import duke.UI;
+import duke.task.Task;
+
 public class DeleteCommand extends Command {
     private final Task task;
     private final int index;
@@ -10,7 +18,7 @@ public class DeleteCommand extends Command {
     @Override
     public void execute(Tasklist tasks, UI ui, FileManager fileManager) throws DukeException {
         tasks.delete(this.index);
-        ui.deleteTask(this.task);
+        ui.deleteTask(tasks.getTask(this.index));
         fileManager.updateTaskList(tasks, ui);
     }
