+++ src/main/java/duke/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.FileManager;
+import duke.Tasklist;
+import duke.UI;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(Tasklist task, UI ui, FileManager fileManager) {
