+++ src/main/java/duke/task/Event.java
@@ -1,32 +1,35 @@
+package duke.task;
+
+import duke.DukeException;
+
 import java.time.DateTimeException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.Arrays;
 
-public class Deadline extends Task {
+public class Event extends Task {
     final private LocalDateTime date;
-    final private static String inputExample = "deadline return book /by 3/4/2021 400";
+    final private static String inputExample = " event my birtday /at 01/01/2000 1400";
 
-    public static Deadline of(String input) throws DukeException {
-        String[] eachWord = input.split("/by");
+    public static Event of(String input) throws DukeException {
+        String[] eachWord = input.split("/at");
         if (eachWord.length == 0 || eachWord[0].length() == 0 || eachWord[0].equals(" ")) {
-            throw new DukeException("Description cannot be empty. Type description before /by\nEg."
-                    + Deadline.inputExample);
+            throw new DukeException("Description cannot be empty. Type description before /at\nEg."
+                    + Event.inputExample);
         }
         if (eachWord.length == 1 || eachWord[1].length() == 0 || eachWord[1].equals(" ")) {
-            throw new DukeException("The date for this event cannot be empty. Type date after /by\nEg."
-                    + Deadline.inputExample);
+            throw new DukeException("The date for this event cannot be empty. Type date after /at\nEg."
+                    + Event.inputExample);
         }
         String dateDescription = eachWord[1];
         String[] dateSplitBySpace = dateDescription.split(" ");
         if (dateSplitBySpace.length < 3) {
             throw new DukeException("Enter the date for this event in DD/MM/YYYY HHMM format\nEg."
-                    + Deadline.inputExample);
+                    + Event.inputExample);
         }
         String[] dateArr = dateSplitBySpace[1].split("/");
         if (dateArr.length < 3) {
             throw new DukeException("Enter the date for this event in DD/MM/YYYY HHMM format\nEg."
-                    + Deadline.inputExample);
+                    + Event.inputExample);
         }
         try {
             int year = Integer.parseInt(dateArr[2]);
@@ -35,24 +38,19 @@ public class Deadline extends Task {
             int hour = Integer.parseInt(dateSplitBySpace[2].substring(0, dateSplitBySpace[2].length() - 2));
             int min = Integer.parseInt(dateSplitBySpace[2].substring(dateSplitBySpace[2].length() - 2));
             LocalDateTime dateTime = LocalDateTime.of(year, month, date, hour, min);
-            return new Deadline(eachWord[0], dateTime);
+            return new Event(eachWord[0], dateTime);
         } catch (NumberFormatException e) {
             throw new DukeException("Enter the date for this event in DD/MM/YYYY HHMM format\nEg."
-                    + Deadline.inputExample);
+                    + Event.inputExample);
         } catch (DateTimeException e) {
             throw new DukeException("Invalid date inputed. Please check that the date is correct\nEg."
-                    + Deadline.inputExample);
+                    + Event.inputExample);
         }
     }
 
-    private Deadline(String description, LocalDateTime dateTime) {
-        super(description);
-        this.date = dateTime;
-    }
-
-//    private Deadline(String input) {
-//        super(input.split("/by", 2)[0]);
-//        String dateDescription = input.split("/by", 2)[1];
+//    private duke.task.Event(String input) {
+//        super(input.split("/at", 2)[0]);
+//        String dateDescription = input.split("/at", 2)[1];
 //        String[] dateSplitBySpace = dateDescription.split(" ");
 //        String[] dateArr = dateSplitBySpace[1].split("/");
 //        int hour = Integer.parseInt(dateSplitBySpace[2].substring(0, dateSplitBySpace[2].length() - 2));
@@ -61,15 +59,20 @@ public class Deadline extends Task {
 //                Integer.parseInt(dateSplitBySpace[2].substring(dateSplitBySpace[2].length() - 2)));
 //    }
 
+    private Event(String description, LocalDateTime dateTime) {
+        super(description);
+        this.date = dateTime;
+    }
+
     @Override
     public String toString() {
-        return ("[D]" + super.toString() + String.format("(by:%s)", this.date));
+        return ("[E]" + super.toString() + String.format("(at:%s)", this.date));
     }
 
     @Override
     public String typeString() {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy HHmm");
         String formatDateTime = this.date.format(formatter);
-        return "deadline" + Task.sep + super.toSaveInFile("/by " + formatDateTime);
+        return "event" + Task.sep + super.toSaveInFile("/at " + formatDateTime);
     }
 }
